lib/column: fix base type for IPv4 and IPv6
Using `net.IP{}` as a base type for an IPv4 is incorrect as it is not a
valid IPv4. This can be verified by calling the `To4()` which should
return non-nil value for an IPv4 address. To address this problem, use
the actual zero value for an IPv4 address (`net.IPv4zero`).

Similarly, use `net.IPv6unspecified` for the IPv6 type.