fix: unsafe conversion from string to []bytes
The previous code was pointing the SliceHeader to the string's header.
This is wrong because they aren't the same size, SliceHeader also
contains the cap. The code was overwriting the cap value to be equal to
the len of str, but this cause issue when str is allocated on the stack.
In that case, it will overwrite the return address of the function.

Instead of fixing the original code, I have copied a well-documented and
tested version from m3db (License Apache 2).