Unshadow url import in open() (bootstrap.go)
open() in bootstrap.go was previously shadowing the `net/url` import, thus
causing the import to be unusable after the first line of the function
(`url, err := url.Parse(dsn)`). This is *generally* bad practice and
should be avoided. A quick renaming of the variable from `url` to
`parsedUrl` allows for usage of the `net/url` import throughout the rest
of open(), and prevents unexpected behavior.