conn: fix Close() on rows when a timeout during processing
There is a race condition when calling rows.Close() after a query was
stopped due to an error (notably when hitting a timeout). The
goroutine processing the query result tries to send the error to the
error channel, before closing both the error and the stream channel
while the Close() method is trying to read from stream before
processing errors. This leads to a deadlock.

This could be fixed by closing the stream channel before sending the
error, but it seems more future-proof to drain both stream and errors
channel in parallel in the Close() method.