Fully parse secure and skip_verify DSN params
The current code will interpret a DSN with `?secure=false` in its
options as a directive to turn TSL _on!_

Obviously re-setting what is the default value is redundant, but
deployment systems that template the DSN from e.g. environment variables
may need to do this.

The previous behavior (`?secure` without a value set turns secure on) is
preserved to minimize surprise.

Signed-off-by: Nathan J. Mehl <n@oden.io>