Handle EPIPE errors
Presently, if the tcp session to the server closes, we will emit broken
pipe errors until either the connection comes back to "idle" state or
until we reach ConnMaxLifetime.

Since EPIPE means that the tcp session is dead (ie we received a RST
packet from the server), there is no point in attempting to proceed past
that point: set the connection as closed.